Switch size_t to int, getting hard to track where -1 may be used - Various cases where -1 is used or >=0 is used, particularly in sortPalettes()

+++ src/quant.h
@@ -10,7 +10,7 @@
 using namespace std;
 
 struct Candidate {
-    size_t    colorIndex;
+    int       colorIndex;
     double    colorDistance;
     rgbColor  comparedColor;
     double    brightness;
